Fixed issue #10. Now builds correctly on windows.
